updates to ARIA table to bring in 1.2 references/allowances
For the cells that are non-normative, we can link to ARIA 1.2.

This will help close some issues, and make sure this spec doesn’t confuse people.

New roles have been added but are commented out until 1.2 is a w3c recommendation.

This initial commit updates the table from `alert` to `menu`.