chore: source cleanup
there have been some markup inconsistencies in the html elements table.  this PR is looking to correct those.
Additionally, this PR moves the IDs from the `tr` element to the `th` for each row.  This is because the `tabindex=-1` is causing chromium to calculate an accessible name for the row based on all the cells, and that's not very pleasant.  Moving to the `th` negates this name calculation while still providing a reasonable place to move focus to for each permalinked row.