add the new search element
keeping this straight forward for the new element's initial inclusion.

being that this is a new element, the reasoning for authors to need to immediately apply overwriting roles is irrelevant.  Thus I see no reason to initially allow any roles on this element.

All global `aria-*` attributes would be allowed on this element.