Addition: update figure role allowances
closes #412

As discussed in https://github.com/w3c/dpub-aria/issues/42#issuecomment-1092558918, `doc-example` is a dpub variant of the `figure` role.  This role being the same as `figure` but with the opportunity of being exposed as an "example" for its role description, needs to be an allowed role for the `figure` element.