Editorial: clarify summary element role mapping
Closes #355

Since the original issue was filed, and I have spent more time researching details/summary and trying to get consensus on what to do with the summary element in particular, I think it does make sense to adjust this per JAWS-test's issue.  The role is now stated as "no corresponding role" since there is no direct ARIA mapping in HTML AAM, but the note was added to indicate the reality of how the role is actually mapped.