Editorial: update and expand section 3
This PR removes the redundant mentions of "this section is non-normative".  We did not need this for each sub-section, since the parent section 3 is already marked as non-normative.

This PR clarifies and modifies some of the original examples used in this section, and adds a completely new sub-section indicating that authors still need to respect the rules of the host language - using an invalid nesting example to demonstrate.