revisions / updates to the PR
incorporates feedback from steve and sarah.  includes language "the following roles are allowed, but are NOT RECOMMENDED" to introduce new allowed roles which _can_ work depending on what's been built, but there are probably better ways to do this.  This wording is also used in #446, and allows for us to indicate these changes per element, without having to restructure the entire table (as this can be the end goal, rather than trying to get all allowed roles for all elements updated at once).