adjust attributes table content.  add command attr rules
the various lines of "authors may use [ attribute name ]" were unnecessary for the scope of this spec.

listing out what elements the HTML attributes are used with was also unnecessary / and would require upkeep that adds extra maintenance to this spec.

adds rules for command attribute, and adds aria-details allowances for popovertarget.