finish command guidance
update the first column which had copied the `contenteditable` content to the correct `command` related content.

update the allowances for aria-expanded to only be applicable to the toggle/show/hide popover states - since that's where the implicit conflict would come into play.

other wording fixes for the attribute table cleanup that i was doing in this PR as well - for better/worse since that arguably should have been another PR... but the work is done so here it is...