Correction for HTML element role mapping (#550)
Correction for HTML element role mapping

per https://github.com/w3c/aria/pull/2504 - the html element doesn't actually map to the `document` role - rather a parent `document node` to the `html` element is what exposes the `document` role of a web page.

This update fixes the spec to match the updated HTML aam mapping / matches reality.  

The `generic` role is being added as an "allowed" but not recommended role to the `html` element.  The `document` role is being kept as an allowed role for the `html` element.  The rational there being that this has been in the spec for a long time and even when specified on the html element, from what i can tell, there's no pro/con to its inclusion.